# ci: VM integration tests speedup