--- a/.github/workflows/pull_request_integration_tests.yml
+++ b/.github/workflows/pull_request_integration_tests.yml
@@ -20,15 +20,6 @@ jobs:
         with:
           persist-credentials: false
 
-      - name: Set up Go
-        uses: actions/setup-go@d35c59abb061a4a6fb18e82ac0862c26744d6ab5 # v5.5.0
-        with:
-          go-version: "1.25"
-          cache: false
-
-      - name: Generate files
-        run: make docker-generate
-
       - name: build matrix
         id: build-matrix
         env:
@@ -66,7 +57,7 @@ jobs:
         run: make prereqs docker-generate
 
       - name: Install gotestsum
-        run: go install gotest.tools/gotestsum@v1.12.3
+        run: make tools
 
       - name: Check disk usage before tests
         run: df -h
@@ -91,7 +82,7 @@ jobs:
 
           mkdir -p /home/runner/reports
 
-          ~/go/bin/gotestsum -ftestname \
+          ./.tools/gotestsum -ftestname \
               --jsonfile=/home/runner/reports/test-run-"${RUN_NUMBER}"-"${MATRIX_ID}".log \
               -- -race -tags=${TEST_TAGS} -timeout 30m \
               -run="^(${MATRIX_TEST_PATTERN})$" ./test/integration/...