--- a/.github/workflows/workflow_integration_tests_vm.yml
+++ b/.github/workflows/workflow_integration_tests_vm.yml
@@ -6,7 +6,7 @@ on:
     paths: ["bpf/**"]
   pull_request:
     branches: ["main", "release-*"]
-    # paths: ["bpf/**"]
+    paths: ["bpf/**"]
 
 permissions:
   contents: read