--- a/internal/tools/tools.go
+++ b/internal/tools/tools.go
@@ -12,6 +12,7 @@ import (
 	_ "github.com/google/go-licenses/v2"
 	_ "github.com/grafana/go-offsets-tracker/cmd/go-offsets-tracker"
 	_ "github.com/onsi/ginkgo/v2/ginkgo"
+	_ "gotest.tools/gotestsum"
 	_ "sigs.k8s.io/controller-runtime/tools/setup-envtest"
 	_ "sigs.k8s.io/kind"
 )