--- a/test/tools/dir.go
+++ b/test/tools/dir.go
@@ -4,6 +4,7 @@
 package tools
 
 import (
+	"os"
 	"path"
 	"path/filepath"
 	"runtime"
@@ -11,6 +12,11 @@ import (
 
 // ProjectDir returns the path of the project's root folder
 func ProjectDir() string {
+	// Check for environment variable override (useful for pre-compiled binaries)
+	if projectDir := os.Getenv("TEST_PROJECT_DIR"); projectDir != "" {
+		return filepath.Clean(projectDir)
+	}
+
 	_, thisFile, _, ok := runtime.Caller(0)
 	if !ok {
 		panic("can't get runtime caller(0) file path")