--- a/test/vm/Makefile
+++ b/test/vm/Makefile
@@ -8,11 +8,15 @@ ROOT_IMG ?= $(WORKDIR)/rootfs.qcow2
 ROOT_IMAGE_RAW ?= $(WORKDIR)/rootfs.raw
 REPO_ROOT ?= ../..
 TARGET_TEST ?= run-integration-test-vm
+TEST_PATTERN ?= TestMultiProcess
+RUN_NUMBER ?= 1
 TEST_OUTPUT = $(REPO_ROOT)/testoutput
 MNT ?= mnt
 IMG_SIZE ?= 10G
 QEMU_BIN = qemu-system-$(ARCH)
 QEMU := $(shell which $(QEMU_BIN))
+PRECOMPILED_TESTS_DIR ?= 
+COMPILED_TEST_BINARY ?= compiled-integration-test
 
 .PHONY: all clean launchvm change-owner check_qemu
 
@@ -30,7 +34,11 @@ $(TEST_OUTPUT):
 
 $(ROOT_IMG): Dockerfile
 	# Build the Docker image and create a tarball
-	DOCKER_BUILDKIT=1 docker build -f Dockerfile --build-arg target=$(TARGET_TEST) --output "type=tar,dest=$(TARBALL)" .
+	DOCKER_BUILDKIT=1 docker build --quiet -f Dockerfile \
+		--build-arg target=$(TARGET_TEST) \
+		--build-arg test_pattern=$(TEST_PATTERN) \
+		--build-arg run_number=$(RUN_NUMBER) \
+		--output "type=tar,dest=$(TARBALL)" .
 	# Create a raw image, format it, and extract the tarball
 	fallocate -l $(IMG_SIZE) $(ROOT_IMAGE_RAW)
 	mkfs.ext4 $(ROOT_IMAGE_RAW)
@@ -43,10 +51,16 @@ $(ROOT_IMG): Dockerfile
 	qemu-img convert $(ROOT_IMAGE_RAW) -O qcow2 $(ROOT_IMG) && rm $(ROOT_IMAGE_RAW)
 
 launchvm: check_qemu $(ROOT_IMG) $(TEST_OUTPUT)
-	qemu-system-$(ARCH) -enable-kvm -cpu host -m 5120 -smp 8 -kernel $(KERNEL) -drive file=$(ROOT_IMG),format=qcow2 \
+	@PRECOMPILED_VIRTFS=""; \
+	if [ -n "$(PRECOMPILED_TESTS_DIR)" ] && [ -d "$(PRECOMPILED_TESTS_DIR)" ]; then \
+		echo "Starting QEMU with pre-compiled tests from $(PRECOMPILED_TESTS_DIR)"; \
+		PRECOMPILED_VIRTFS="-virtfs local,path=$(PRECOMPILED_TESTS_DIR),mount_tag=precompiled,security_model=mapped,id=precompiled"; \
+	fi; \
+	qemu-system-$(ARCH) -enable-kvm -cpu host -m 8192 -smp 3 -kernel $(KERNEL) -drive file=$(ROOT_IMG),format=qcow2 \
 		-append "earlyprintk=ttyS0 console=ttyS0 root=/dev/sda rw quiet" \
 		-virtfs local,path=$(REPO_ROOT),mount_tag=beyla,security_model=mapped,id=beyla \
 		-virtfs local,path=$(TEST_OUTPUT),mount_tag=testout,security_model=mapped,id=testout \
+		$$PRECOMPILED_VIRTFS \
 		-net user,hostfwd=tcp::2222-:22 -net nic \
 		-nographic